Tidy up the notice migration
